server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy>=2.0.36
pymysql==1.1.0
cryptography==42.0.5
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.9
python-dotenv==1.0.1
pillow>=10.4.0
aiofiles==23.2.1
